Changed control panel ui, support for view/edit mode in topology editor

+++ webservice/src/main/resources/assets/scripts/modules/TopologyGraphCreator.js
@@ -126,8 +126,10 @@ define(['require',
 				};
 			})
 			.on("drag", function(args) {
-				thisGraph.state.justDragged = true;
-				thisGraph.dragmove.call(thisGraph, args);
+				if(thisGraph.editMode) {
+					thisGraph.state.justDragged = true;
+					thisGraph.dragmove.call(thisGraph, args);
+				}
 			})
 			.on("dragend", function() {
 				// todo check if edge-mode is selected
@@ -144,8 +146,8 @@ define(['require',
 			thisGraph.svgMouseDown.call(thisGraph, d);
 		});
 		svg.on("mouseup", function(d) {
-			thisGraph.hideShuffle();
 			thisGraph.svgMouseUp.call(thisGraph, d);
+			// thisGraph.hideShuffle();
 		});
 
 		// listen for dragging
@@ -313,7 +315,6 @@ define(['require',
 			});
 
 		for (var i = 0; i < words.length; i++) {
-			// var tspan = el.append('tspan').text(words[i]);
 			nodeTitle += words[i]+' ';
 		}
 		el.text(nodeTitle.trim());
@@ -827,24 +828,33 @@ define(['require',
 			.attr("data-toggle", "popover")
 			.attr("data-name", function(d){ return d.source.uiname +'-'+d.target.uiname; })
 			.on("mouseover", function(d){
-				// $('[data-uiname="'+d.source.uiname+'-'+d.target.uiname+'"]').show();
+				if(thisGraph.editMode) {
+					// $('[data-uiname="'+d.source.uiname+'-'+d.target.uiname+'"]').show();
+				}
 			})
 			.on("mouseout", function(d){
-				// $('[data-uiname="'+d.source.uiname+'-'+d.target.uiname+'"]').hide();
+				if(thisGraph.editMode) {
+					// $('[data-uiname="'+d.source.uiname+'-'+d.target.uiname+'"]').hide();
+				}
 			})
 			.on("mousedown", function(d) {
-				if(d3.event.shiftKey){
-					var elem = $(this).parent().find('.visible-link[d="'+$(this).attr("d")+'"]')[0];
-					thisGraph.pathMouseDown.call(thisGraph, d3.select(elem), d);
-				} else {
-					// thisGraph.showShuffle(d);
+				if(thisGraph.editMode) {
+					if(d3.event.shiftKey){
+						var elem = $(this).parent().find('.visible-link[d="'+$(this).attr("d")+'"]')[0];
+						thisGraph.pathMouseDown.call(thisGraph, d3.select(elem), d);
+					} else {
+						// thisGraph.showShuffle(d);
+					}
 				}
 			})
 			.on("mouseup", function(d) {
-				if(d3.event.shiftKey){
-					state.mouseDownLink = null;
-				} else {
-					// thisGraph.showShuffle(d);
+				if(thisGraph.editMode){
+					if(d3.event.shiftKey){
+						state.mouseDownLink = null;
+					} else {
+						// $('[data-uiname="'+d.source.uiname+'-'+d.target.uiname+'"][data-toggle="popover"]').popover('show');
+						// thisGraph.bindPopoverDropdown(d);
+					}
 				}
 			});
 
@@ -865,17 +875,34 @@ define(['require',
             .attr("data-toggle", "popover")
             .attr("data-type", "shuffle")
 			.on('mouseover', function(d){
-				// $(this).show();
+				if(thisGraph.editMode) {
+					// $(this).show();
+				}
 		    })
 			.on('mouseout', function(d){
-				// $(this).hide();
+				if(thisGraph.editMode) {
+					// $(this).hide();
+				}
 		    })
 		    .on("mousedown", function(d) {
-		    	// thisGraph.showShuffle(d);
+		    	if(thisGraph.editMode) {
+		    		if(d3.event.shiftKey){
+						var elem = $(this).parent().find('.visible-link[d="'+$(this).attr("d")+'"]')[0];
+						thisGraph.pathMouseDown.call(thisGraph, d3.select(elem), d);
+					} else {
+						// thisGraph.showShuffle(d);
+					}
+		    	}
 			})
 			.on("mouseup", function(d) {
-				// thisGraph.showShuffle(d);
-				// state.mouseDownLink = null;
+				if(thisGraph.editMode) {
+					if(d3.event.shiftKey){
+						state.mouseDownLink = null;
+					} else {
+						// $('[data-uiname="'+d.source.uiname+'-'+d.target.uiname+'"][data-toggle="popover"]').popover('show');
+						// thisGraph.bindPopoverDropdown(d);
+					}
+				}
 			});
 
 		// remove old links
@@ -883,6 +910,9 @@ define(['require',
 
 		//set shuffle icon on links
 		this.setLinkIcon();
+
+		//adding dropdown for toggle
+		this.showShuffle();
 
 		//clone the paths or links to make hover on them with some hidden margin
 		thisGraph.clonePaths();
@@ -930,12 +960,16 @@ define(['require',
 				.attr("width", "48px")
 				.attr("height", "48px")
 			    .on("mouseover", function(d) {
-			    	$(this).css("opacity", "0.75");
-					$(this).siblings('text.fa-times').show();
+			    	if(thisGraph.editMode){
+			    		$(this).css("opacity", "0.75");
+						$(this).siblings('text.fa-times').show();
+			    	}
 				})
 				.on("mouseout", function(d) {
-					$(this).css("opacity", "1");
-					$(this).siblings('text.fa-times').hide();
+					if(thisGraph.editMode){
+						$(this).css("opacity", "1");
+						$(this).siblings('text.fa-times').hide();
+					}
 				})
 				.on("mousedown", function(d) {
 					thisGraph.rectangleMouseDown.call(thisGraph, d3.select(this.parentNode), d);
@@ -955,13 +989,19 @@ define(['require',
                 .style("display","none")
                 .style("font-size","large")
 				.on('mouseover', function(d){
-					$(this).show();
+					if(thisGraph.editMode){
+						$(this).show();
+					}
 			    })
 				.on('mouseout', function(d){
-					$(this).hide();
+					if(thisGraph.editMode){
+						$(this).hide();
+					}
 			    })
 				.on('mousedown', function(d){
-					thisGraph.deleteNode(d);
+					if(thisGraph.editMode){
+						thisGraph.deleteNode(d);
+					}
 			    });
 
 			newGs.append("circle")
@@ -1006,10 +1046,14 @@ define(['require',
 					d3.select(this).classed(consts.connectClass, false);
 				})
 				.on("mousedown", function(d) {
-					thisGraph.circleMouseDown.call(thisGraph, d3.select(this), d);
+					if(thisGraph.editMode){
+						thisGraph.circleMouseDown.call(thisGraph, d3.select(this), d);
+					}
 				})
 				.on("mouseup", function(d) {
-					thisGraph.circleMouseUp.call(thisGraph, d3.select(this), d);
+					if(thisGraph.editMode){
+						thisGraph.circleMouseUp.call(thisGraph, d3.select(this), d);
+					}
 				})
 				.call(thisGraph.drag);
 
@@ -1048,10 +1092,14 @@ define(['require',
 					d3.select(this).classed(consts.connectClass, false);
 				})
 				.on("mousedown", function(d) {
-					thisGraph.circleMouseDown.call(thisGraph, d3.select(this), d);
+					if(thisGraph.editMode){
+						thisGraph.circleMouseDown.call(thisGraph, d3.select(this), d);
+					}
 				})
 				.on("mouseup", function(d) {
-					thisGraph.circleMouseUp.call(thisGraph, d3.select(this), d);
+					if(thisGraph.editMode){
+						thisGraph.circleMouseUp.call(thisGraph, d3.select(this), d);
+					}
 				})
 				.call(thisGraph.drag);
 
@@ -1079,7 +1127,9 @@ define(['require',
 					} else return "";
 				})
 		        .on("mouseup", function(d) {
-					thisGraph.circleMouseUp.call(thisGraph, d3.select(this), d);
+		        	if(thisGraph.editMode){
+		        		thisGraph.circleMouseUp.call(thisGraph, d3.select(this), d);
+		        	}
 				});
 		        
 		        
@@ -1093,15 +1143,13 @@ define(['require',
 		thisGraph.rectangles.exit().remove();
 	};
 
-	TopologyGraphCreator.prototype.showShuffle = function(d){
+	TopologyGraphCreator.prototype.showShuffle = function(){
 		var thisGraph = this;
 		var shuffleArr = thisGraph.data.linkShuffleOptions;
 		var html = "<div><select class='link-shuffle'>";
 		for(var i = 0, len = shuffleArr.length; i < len; i++){
 			html += "<option value='"+shuffleArr[i].val+"'>"+shuffleArr[i].label+"</option>";
 		}
-		//<div class='link-btns'><button class='btn btn-xs btn-default'><i class='fa fa-times'></i></button>"+
-				// "<button class='btn btn-xs btn-success'><i class='fa fa-check'></i></button></div>
 		html += "</select></div>";
 		$('[data-toggle="popover"][data-type="shuffle"]').popover({
 			title: "Select Grouping",
@@ -1114,7 +1162,19 @@ define(['require',
 	};
 
 	TopologyGraphCreator.prototype.hideShuffle = function(){
+		$('.link-shuffle').off('change');
 		$('[data-toggle="popover"][data-type="shuffle"]').popover('hide');
+	};
+
+	TopologyGraphCreator.prototype.bindPopoverDropdown = function(d){
+		$('.link-shuffle').on('change', function(e){
+    		console.log(d);
+    		console.log(e);
+    	});
+	};
+
+	TopologyGraphCreator.prototype.unbindPopoverDropdown = function(){
+		$('.link-shuffle').off('change');
 	};
 
 	TopologyGraphCreator.prototype.clonePaths = function(){
